sendStaion delete btn

+++ web/src/main/webapp/WEB-INF/waybill/send_nextStation/list.jsp
@@ -127,11 +127,15 @@
                 } else if (obj.event === 'tool-edit') {
                     edit(handleNo);
                 }else if (obj.event === 'tool-delete') {
-                    if(data.status==1){
-                        layer.msg("Has been shipped", {icon: 2, time: 2000});
-                        return;
-                    }
-                    deleteHandle(handleNo);
+                    layer.confirm('Are you sure？', {
+                        btn: ['Yes','No'] //按钮
+                    }, function(){
+                        if(data.status==1){
+                            layer.msg("Has been shipped", {icon: 2, time: 2000});
+                            return;
+                        }
+                        deleteHandle(handleNo);
+                    });
                 }
 
             });
